+++ js/main.js
@@ -1,4 +1,13 @@
 import {showScreen} from './utils';
-import getGreetingScreen from './greeting/get-greeting-screen';
+import GreetingScreenView from './greeting/greeting-sreen-view';
+import {initialState, currentState, questions} from './data/game-data';
+import {showGameScreen} from './change-screen/change-screen';
 
-showScreen(getGreetingScreen());
+const greetingScreen = new GreetingScreenView();
+greetingScreen.startGame = () => {
+  Object.assign(currentState, initialState, {
+    answers: []
+  });
+  showGameScreen(currentState, questions);
+};
+showScreen(greetingScreen.element);
